KAFKA-17840: Move QuotaFactory to server module
