+++ core/src/main/java/kafka/server/builders/KafkaApisBuilder.java
@@ -21,7 +21,6 @@ import kafka.coordinator.transaction.TransactionCoordinator;
 import kafka.network.RequestChannel;
 import kafka.server.KafkaApis;
 import kafka.server.KafkaConfig;
-import kafka.server.QuotaFactory.QuotaManagers;
 import kafka.server.ReplicaManager;
 import kafka.server.share.SharePartitionManager;
 
@@ -40,6 +39,7 @@ import org.apache.kafka.server.ClientMetricsManager;
 import org.apache.kafka.server.FetchManager;
 import org.apache.kafka.server.ForwardingManager;
 import org.apache.kafka.server.authorizer.Authorizer;
+import org.apache.kafka.server.quota.QuotaFactory.QuotaManagers;
 import org.apache.kafka.storage.log.metrics.BrokerTopicStats;
 
 import java.util.Map;
