+++ core/src/main/java/kafka/server/share/ReplicaManagerLogReader.java
@@ -16,13 +16,13 @@
  */
 package kafka.server.share;
 
-import kafka.server.QuotaFactory;
 import kafka.server.ReplicaManager;
 
 import org.apache.kafka.common.TopicIdPartition;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.FetchRequest;
 import org.apache.kafka.server.log.remote.storage.RemoteLogManager;
+import org.apache.kafka.server.quota.QuotaFactory;
 import org.apache.kafka.server.share.LogReader;
 import org.apache.kafka.server.storage.log.FetchParams;
 import org.apache.kafka.storage.internals.log.FetchDataInfo;
