+++ core/src/main/scala/kafka/server/metadata/ClientQuotaMetadataManager.scala
@@ -18,7 +18,7 @@
 package kafka.server.metadata
 
 import kafka.network.ConnectionQuotas
-import kafka.server.QuotaFactory.QuotaManagers
+import org.apache.kafka.server.quota.QuotaFactory.QuotaManagers
 import kafka.server.metadata.ClientQuotaMetadataManager.transferToClientQuotaEntity
 import kafka.utils.Logging
 import org.apache.kafka.common.metrics.Quota
