+++ core/src/test/java/kafka/server/share/DelayedShareFetchTest.java
@@ -17,7 +17,6 @@
 package kafka.server.share;
 
 import kafka.cluster.Partition;
-import kafka.server.QuotaFactory;
 import kafka.server.ReplicaManager;
 
 import org.apache.kafka.clients.consumer.internals.ShareAcquireMode;
@@ -35,6 +34,7 @@ import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.server.log.remote.storage.RemoteLogManager;
 import org.apache.kafka.server.purgatory.DelayedOperationKey;
 import org.apache.kafka.server.purgatory.DelayedOperationPurgatory;
+import org.apache.kafka.server.quota.QuotaFactory;
 import org.apache.kafka.server.quota.ReplicaQuota;
 import org.apache.kafka.server.share.LogReader;
 import org.apache.kafka.server.share.PartitionMetadataProvider;
