+++ core/src/test/scala/unit/kafka/server/ReplicaAlterLogDirsThreadTest.scala
@@ -17,7 +17,7 @@
 package kafka.server
 
 import kafka.cluster.Partition
-import kafka.server.QuotaFactory.UNBOUNDED_QUOTA
+import org.apache.kafka.server.quota.QuotaFactory.UNBOUNDED_QUOTA
 import kafka.server.ReplicaAlterLogDirsThread.ReassignmentState
 import kafka.utils.TestUtils
 import org.apache.kafka.common.errors.KafkaStorageException
